perf(core): optimize BHVec10240 bundling via 100% safe transposed bit-sliced addition
